backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
python-multipart==0.0.20
SQLAlchemy==2.0.41
openai==1.70.0
faster-whisper==1.1.1
ffmpeg-python==0.2.0
pydantic==2.11.2
orjson==3.10.16
requests==2.32.3

